agent3/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langgraph
langchain-google-genai
httpx>=0.25.0
pydantic>=2.0.0
a2a-sdk[http-server]
